Move UnqualifiedName to Query and remove SplitByIndicies
